refactor: split map view into modules
